% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BornOpenData.eprime2.R
\docType{package}
\name{BornOpenData.eprime2-package}
\alias{BornOpenData.eprime2}
\alias{BornOpenData.eprime2-package}
\title{BornOpenData.eprime2}
\description{
This package has a function to read data stored as BornOpenData by E-Prime
when complemented with the package BornOpenData4EPrime.
}
\details{
BornOpenData.eprime2 is a package with a single function, 
readBornOpenData.eprime2, meant to get data from an E-Prime 
experiment made BornOpenData with the package BornOpenData4EPrime. 

readBornOpenData.eprime2(owner, repository, verbose = FALSE )  to read all the data on the repository.

where owner, repository points to a git hub repository that can be consulted at
    https://github.com/owner/repository
and the data files themselves are located in
    https://github.com/owner/repository/rawdata

See help(readBornOpenData.eprime2) for the function.
}
\keyword{internal}
