% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBornOpenData.eprime2.R
\name{readBornOpenData.eprime2}
\alias{readBornOpenData.eprime2}
\title{readBornOpenData.eprime2}
\usage{
readBornOpenData.eprime2(owner, repository, verbose = FALSE)
}
\arguments{
\item{owner}{is the GitHub member who created the repository}

\item{repository}{is the name of the repository on GitHub}

\item{verbose}{= TRUE displays complementary information.}
}
\value{
the data of all participants, all sessions in a data.frame
}
\description{
This function reads data from E-Prime stored on GitHub
by the E-Prime package BornOpenData4Eprime. Note that the function
requires 10-30 seconds to retrive the information from GitHub. Also,
an internet connection must be effective.
}
\details{

}
\examples{
# Assume a user on GitHub called VIC-Laboratory-ExperimentalData
# and a repository created by that user called Test.
# You can check online at https://github.com/VIC-Laboratory-ExperimentalData/Test
# that this user and this repository exists.

# This will read all the data files (stored in the rawdata folder):
dta <- readBornOpenData.eprime2("VIC-Laboratory-ExperimentalData","Test")

}
\author{
Denis Cousineau, \email{denis.cousineau@uottawa.ca}
}
\keyword{Born}
\keyword{Data,}
\keyword{E-Prime}
\keyword{Open}
